requirements.txt
gradio==6.0.2
requests==2.32.5
python-dotenv==1.2.1
semanticscholar==0.11.0
PyMuPDF==1.26.6
huggingface_hub==0.23.0
nltk==3.9.2
rouge-score==0.1.2

